small JS fixes
`New Array(len)` doesn't actually do what it does in most languages in JS. It does not allocate an array of a specific size, it simply creates an array object and sets its length property to the indicated value, similar to doing `arr = []; arr.length = 10`. It was added way back when `document.write` was still a good idea, and never got rescinded.

One thing you can use it for is trick code, like `weights = new Array(n+1).join(',').split(',').map(v => 1)`, which builds a new array without allocation, then takes advantage of the iteration process (which relies purely on `length`) to join up the array into a string `,,,,,,,` of length `n`, then splits it up again to form a *real* array ['','','','',...] of length `n`, and then maps that array to some value set. In the preceding code, an array of ones of length `n`, but it's far less readable, and far more expensive, than a simple for loop =)

I've also made some of the variables that are function scoped explicitly function scoped, because code in strict mode will throw errors over those repeated `var i` etc. declarations in the for loops. `var` creates a function-scoped variable, so two consecutive `for (var i=0; ...)` will lead to a redeclaration error (along the lines of "variable already exists"). ES6 has `let` these days, but there are still plenty of systems that use a Node that's older than v4, so updating the code to ES6 syntax seems premature at best, and a breaking change on average.

Also you're my hero of the week, it is miraculous how hard it is to find an actual complete, generic implementation of BSplines, let along in JavaScript!